# src/server/swagger/swagger.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Network Detection and Response (NDR)",
    "description" : "Advanced Cybersecurity Analytics Service (ACAS) for AI-based analysis with explainable intelligence of network events captured by the Montimage Monitoring Tools (MMT)",
    "termsOfService" : "https://github.com/Montimage/acas",
    "contact" : {
      "email" : "developer@montimage.com"
    },
    "license" : {
      "name" : "Montimage",
      "url" : "https://www.montimage.eu"
    },
    "version" : "v1"
  },
...
  "servers": [
    {
      "description": "Localhost Server",
      "url": "http://localhost:31057/api/"
    },
    {
      "description": "Production Server",
      "url": "https://ndr.montimage.eu/api/"
    }
  ],
...
  "paths" : {
...
    "/predict/offline" : {
      "post" : {
        "tags" : [ "predictions" ],
        "summary" : "Offline prediction with automatic MMT analysis",
        "description" : "Start offline prediction on a PCAP file. Server automatically handles MMT analysis.",
        "operationId" : "predictOffline",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : ["modelId", "pcapFile"],
                "properties" : {
                  "modelId" : {
                    "type" : "string",
                    "description" : "ID of the trained model to use for prediction",
                    "example" : "model_2"
                  },
                  "pcapFile" : {
                    "type" : "string",
                    "description" : "Name of the uploaded PCAP file. Server handles MMT analysis automatically.",
                    "example" : "infiltration.pcap"
                  },
                  "useQueue" : {
                    "type" : "boolean",
                    "description" : "Use job queue for non-blocking execution (default: true)",
                    "example" : true
                  },
                  "filterIPs" : {
                    "type" : "array",
                    "items" : { "type" : "string" },
                    "description" : "ISIM integration: Optional list of IP addresses to filter. Only traffic involving these IPs will be analyzed.",
                    "example" : ["192.168.1.100", "10.0.0.5"]
                  }
                }
              },
              "example" : {
                "modelId" : "model_2",
                "pcapFile" : "infiltration.pcap",
                "useQueue" : true,
                "filterIPs" : ["192.168.1.100", "10.0.0.5"]
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {
                      "type" : "boolean"
                    },
                    "jobId" : {
                      "type" : "string",
                      "description" : "Job ID for polling status (when useQueue=true)"
                    },
                    "predictionId" : {
                      "type" : "string",
                      "description" : "Prediction ID for retrieving results"
                    },
                    "queueName" : {
                      "type" : "string"
                    },
                    "message" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Missing required parameters or MMT analysis failed"
          },
          "500" : {
            "description" : "Internal server error or MMT timeout"
          }
        }
      }
    },
    "/predict/job/{jobId}" : {
      "get" : {
        "tags" : [ "predictions" ],
        "summary" : "Get queued prediction job status",
        "description" : "Get the status of a queued prediction job",
        "operationId" : "getPredictionJobStatus",
        "parameters" : [ {
          "name" : "jobId",
          "in" : "path",
          "description" : "Job ID",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/jobStatus"
                }
              }
            }
          }
        }
      }
    },
    "/predict/online" : {
      "post" : {
        "tags" : [ "predictions" ],
        "summary" : "Start integrated online prediction with MMT-probe",
        "description" : "Start real-time AI-based intrusion detection on a network interface. This endpoint combines MMT-probe packet capture, DPI analysis, feature extraction, and AI prediction in one integrated pipeline.",
        "operationId" : "startOnlinePrediction",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "required" : ["modelId", "interface"],
                "properties" : {
                  "modelId" : {
                    "type" : "string",
                    "description" : "Model ID to use for prediction",
                    "example" : "model_1"
                  },
                  "interface" : {
                    "type" : "string",
                    "description" : "Network interface to monitor (e.g., eth0, lo)",
                    "example" : "eth0"
                  },
                  "filterIPs" : {
                    "type" : "array",
                    "items" : { "type" : "string" },
                    "description" : "ISIM integration: Optional list of IP addresses to filter. Only traffic involving these IPs will be analyzed. Use with port mirroring for monitoring critical assets.",
                    "example" : ["192.168.1.100", "10.0.0.5"]
                  }
                }
              },
              "example" : {
                "modelId" : "model_1",
                "interface" : "eth0",
                "filterIPs" : ["192.168.1.100", "10.0.0.5"]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Online prediction started successfully",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "success" : {"type" : "boolean"},
                    "mode" : {"type" : "string", "example" : "online"},
                    "interface" : {"type" : "string"},
                    "modelId" : {"type" : "string"},
                    "predictionId" : {"type" : "string"},
                    "isRunning" : {"type" : "boolean"},
                    "startedAt" : {"type" : "number"},
                    "message" : {"type" : "string"}
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Missing required parameters"
          },
          "500" : {
            "description" : "Failed to start online prediction"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "jobStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["completed", "failed", "processing", "waiting"],
            "example": "completed"
          },
          "result": {
            "type": "object",
            "description": "Job result data when completed"
          },
          "failedReason": {
            "type": "string",
            "description": "Error message if job failed"
          },
          "progress": {
            "type": "number",
            "description": "Job progress percentage (0-100)"
          }
        }
      }
    }
  }
}